messagehandler functions are exported to utils to achieve reusablility
